% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articleTable-package.R
\docType{package}
\name{articleTable-package}
\alias{articleTable}
\alias{articleTable-package}
\title{articleTable: Análise do modelo de regressão Chen}
\description{
Considerando o modelo de regressão quantílico proposto por Souza(2021) este pacote, composto por apenas uma função, busca um modo prático de análisar as propriedades dos estimadores de máxima verossimilhança e transformá-las em Latex.
}
\keyword{internal}
