% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Avaliação do modelo de regressao quantílico Chen}
\usage{
make_table(n, theta)
}
\arguments{
\item{n}{Inteiro que representa o número de valores aleatorios a serem gerados
no cálculo das propriedades do estimador.}

\item{theta}{Vetor de numéricos com os parâmetros. O primeiro é considerado lambda,
todos os outros são considerados parâmetros do modelo de regressão quantílico Chen.}
}
\value{
Uma tabela formatada em Latex utilizando o pacote kable. Nela constamas seguintes propriedades do estimador:
Média, Viés; Erro padrão; Erro quadrado médio; Limite inferior do intervalo de
confiança; Limite superior do intervalo de confiança; Taxa de cobertura.
}
\description{
Considerando o modelo de regressão quantílico proposto por Souza(2021) esta função
busca um modo prático de análisar as propriedades dos estimadores de máxima
verossimilhança e transformá-las em Latex.
}
\examples{
theta1 <- c(0.7, 2, -1, 0.8)
make_table(30, theta1)
make_table(100, theta1)
make_table(200, theta1)

theta2 <- c(0.7, 1.5, -0.8, 0.5, 2, -1)
make_table(30, theta2)
make_table(100, theta2)
make_table(200, theta2)
}
